redesign the astrodao vote func add new command /checkproposal
